Refactor resource loader and world mesh builder in SchematicRenderer
